api/requirements.txt
fastapi
uvicorn[standard]
pymongo
qdrant-client
requests
sentence-transformers
PyJWT
passlib[bcrypt]
